Modals display correct information only to admins.
- Creating a new AdminAssingRolForm that reduces the number of roles an admin
can assing from 5 to 4, everything but admin.
- Changed the way Rol are displayed, they now display like this:
<id_number>: <Rol_nombre>
- Adding new function to usuario_asovac Model that returns the biggest or
more important role that the user has so if he has 2, the one with lesser id will
be returned.
- Creating a new ajax modal template to only display the roles of a selected user from
the users list, this will be displayed instead of a form when the current requesting user
does not have a bigger role that the targeted user to change.