Refactoring of sidebar menu options context functions
- Refactoring validate_rol_status(...) in main_app/views.py to fill
a list with the strings of the option names that are suposed to appear
on the frontend, then this list is added to the items dictionary.
- The conditional statements of the sidebar menu options were directly moved
to the inside of validate_rol_status(...).
- Deleting a lot of now unnecesary items from the context dictionary of
each view of the 'main_app' app for better readbility.
- Refactoring the sidebar.html template to display each option only if its
designed string appears on the items.sidebar_options list for a more
readable code.
- Removing unnecessary whitespace.
- NOTE!: There is still many references to the modified functions and
context dictionary changes to be done in the rest of the apps besides
main_app.